Rename : 구조할당분해 컴포넌트 import  재설정
