+++ src/Components/Badge/index.js
@@ -1,1 +1,1 @@
-export { default as Badge } from "./Badge";
+export { default } from "./Badge";
