+++ src/Components/Button/index.js
@@ -1,1 +1,1 @@
-export { default as Button } from "./Button";
+export { default } from "./Button";
