+++ src/Components/ImageBox/ImageBox.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import styled, { css } from "styled-components";
 
-import { Logo } from "../../Basic/Logo";
+import { Logo } from "../../Basic";
 import { glowLight, glowDark } from "../../shared/animation";
 import { color } from "../../shared/styles";
