+++ src/Components/ImageBox/index.js
@@ -1,1 +1,1 @@
-export { default as ImageBox } from "./ImageBox";
+export { default } from "./ImageBox";
