+++ src/Components/index.js
@@ -1,4 +1,4 @@
-// export { default as Badge } from "./Badge";
+export { default as Badge } from "./Badge";
 // export { default as Button } from "./Button";
 export { default as Comment } from "./Comment";
 // export { default as ImageBox } from "./ImageBox";
