+++ src/Layouts/Card/index.js
@@ -1,2 +1,1 @@
-export { default as Card } from "./Card";
-export { default as CardSmall } from "./CardSmall";
+export { default } from "./Card";
