+++ src/Layouts/Sidebar/index.js
@@ -1,1 +1,1 @@
-export { default as SideBar } from "./SideBar";
+export { default } from "./SideBar";
