+++ src/Layouts/SubNavbar/SubNavbar.jsx
@@ -2,7 +2,7 @@ import React, { useState } from "react";
 import PropTypes from "prop-types";
 import styled from "styled-components";
 
-import { SubNavbarItem } from "../../Components/SubNavbarItem";
+import { SubNavbarItem } from "../../Components";
 
 const THEME = {
   LIGHT: "light",
